tests: migrate to pytest
