+++ .gitignore
@@ -4,3 +4,6 @@ __pycache__/
 /*.egg*/
 /build/
 /dist/
+/.coverage*
+/coverage.xml
+/.pytest_cache/
